feat: more cleanup
