+++ apps/torus-portal/src/app/(pages)/permissions/create-permission/capability/_components/create-capability-fields/selected-paths-display.tsx
