+++ apps/torus-portal/src/app/(pages)/permissions/create-permission/capability/_components/create-capability-flow/create-capability-flow-hooks/use-permission-selection.ts
@@ -7,7 +7,7 @@ import type {
   PermissionId,
 } from "@torus-network/sdk/chain";
 
-import type { NamespacePathNodeData } from "../types";
+import type { NamespacePathNodeData } from "../create-capability-flow-types";
 
 interface UsePermissionSelectionProps {
   nodes: Node<NamespacePathNodeData>[];
