+++ apps/torus-portal/src/app/(pages)/permissions/create-permission/capability/_components/create-capability-flow/create-capability-flow-panels/selected-paths-panel.tsx
@@ -1,6 +1,6 @@
 import type { Node } from "@xyflow/react";
 
-import type { NamespacePathNodeData } from "../types";
+import type { NamespacePathNodeData } from "../create-capability-flow-types";
 
 interface SelectedPathsPanelProps {
   rootSelectedPaths: Set<string>;
