+++ apps/torus-portal/src/app/(pages)/permissions/create-permission/capability/_components/create-capability-permission-form-utils.ts
@@ -6,7 +6,7 @@ import type {
 } from "@torus-network/sdk/chain";
 import type { SS58Address } from "@torus-network/sdk/types";
 
-import type { PathWithPermission } from "./create-capability-flow/types";
+import type { PathWithPermission } from "./create-capability-flow/create-capability-flow-types";
 import type { CreateCapabilityPermissionFormData } from "./create-capability-permission-form-schema";
 
 export function transformFormDataToSDK(
